requirements.txt
asgiref
crispy-bootstrap5
distlib
dj-database-url
Django
django-crispy-forms
django-email
django-environ
django-filter
djangorestframework
filelock
Markdown
MouseInfo
pillow
platformdirs
psycopg2
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
PyScreeze
pytweening
sqlparse
typing_extensions
tzdata
virtualenv
